adjusted for extra temperature measurements and filtering for cpu when looking for benchmark

+++ monitoring/utils/console.cpp
@@ -12,7 +12,7 @@ const std::string currentDateTime() {
     return buf;
 }
 
-const int currentDateTimeMilliseconds() {
+const unsigned long long currentDateTimeMilliseconds() {
 	using namespace std::chrono;
 	return duration_cast<milliseconds>(system_clock::now().time_since_epoch()).count();
 }
